requirements1.txt
flask
gtts
gunicorn
google-generativeai
langchain